Rename HudsonPluginDiscoveryPlugin
After some thought it seems simplest to break the most complicated parts
of manifest generation into their own tasks and later combine them.
Renaming since this approach is more about the manifest than the legacy
entry of Plugin-Class.

JENKINS-58205
