+++ src/test/groovy/org/jenkinsci/gradle/plugins/jpi/JpiIntegrationSpec.groovy
@@ -208,8 +208,8 @@ class JpiIntegrationSpec extends IntegrationSpec {
         'check'                                      | ':checkAccessModifier'                        | TaskOutcome.SUCCESS
         'checkAccessModifier'                        | ':compileJava'                                | TaskOutcome.NO_SOURCE
         'checkAccessModifier'                        | ':compileGroovy'                              | TaskOutcome.NO_SOURCE
-        'discoverHudsonPlugins'                      | ':compileJava'                                | TaskOutcome.NO_SOURCE
-        'discoverHudsonPlugins'                      | ':compileGroovy'                              | TaskOutcome.NO_SOURCE
+        'generateJenkinsPluginClassManifest'         | ':compileJava'                                | TaskOutcome.NO_SOURCE
+        'generateJenkinsPluginClassManifest'         | ':compileGroovy'                              | TaskOutcome.NO_SOURCE
     }
 
     @Unroll
